+++ commands/fun/binary.js
@@ -8,7 +8,6 @@ module.exports = {
   run: async (bot, message, args) => {
 
     const Discord = require("discord.js");
-    if (!message.guild) return;
 
     var output = "";
     var input = args.join(" ");