+++ main/functions/timeoutUtils.js
@@ -98,8 +98,8 @@ class TimeoutUtils {
 
                     let present = moment();
                     let untilExpiry = moment.duration((timeout.expires - present.unix()) * 1000);
-                    resolve(untilExpiry.humanize(true))
-                    // resolve(untilExpiry)
+                    // resolve(untilExpiry.humanize(true))
+                    resolve(`${untilExpiry.hours()}h ${untilExpiry.minutes()}m ${untilExpiry.seconds()}s`);
 
                 } else { resolve(false) }
             } catch (error) { reject(error) }